Create form for adding books and remove addBookForm.js
